Allow zero-costs
Currently, the `parepareItem` method checks to see if the cost and notes evaluate to false. If so, it checks from the product list.

However, we need the ability to set cost to zero explicitly through the API. This allows a user to pass null, which will use the default. Otherwise, passing 0 will persist.